HistoryItem interface moved
