--- interfaces/history-item.interface.ts
@@ -1,4 +1,0 @@
-export interface HistoryItem {
-  vhaFilePath: string;
-  hubName: string;
-}
